Refactor env.Environment from wrapping a map to being a map
The main reason behind this refactor is the `*Environment.ToMap()` method. The implication behind the naming of this method is that it returns a copy of the environment in a map representation, but what it actually does is just return the underlying map that Environment is implemented on top of. This is used in a variety of ways across the codebase, and we actually reach into Environment's pockets andmodify this underlying map a fair bit. This in mind, so it makes sense (to me, anyway) to remove the confusingly-named middleman.

One possible drawback of this approach is that it makes it easier to accidentally "manually" set a k/v pair in th env, ie by calling `env["foo"] = "bar"`, buuuuuuuuuuuut it was already pretty easy by calling ToMap(), so i'm happy with this tradeoff, but I'd be very interested in review commets about this.