Unescape backticks when parsing env from export -p output
When we run hooks as part of a build, the agent pulls [some malarkey](https://github.com/buildkite/agent/blob/1a9c92c951fe9c962f4b0be4229f1eeeb6faf35d/hook/scriptwrapper.go) to capture the environment variables added by that hook.

As part of this process, the agent uses the bash `export -p` builtin to export the environment of the current process in a form that can be run as a script. The output looks kinda like:
```Bash
declare -x COLOUR="chartreuse"
declare -x FLAVOUR="spicy"
declare -x TEXTURE="abrasive"
\# ... and so on and so forth
```

We then parse this output into our `env.Environment` type, and use that to construct a diff between what the env looked like before the hook was run, and what it looked like afterward.

The issue here was that this parsing process didn't know how to (or that it should) unescape a backtick (\`). In practice, this means that if a user exports an environment variable with a backtick in it, then the variable will come out with an extra backslash character in it.

For example, if i had an environment hook like:
```Bash
export MARKDOWN_TEXT="this should be formatted like \`code\`"
```
then i would expect to be that `echo $MARKDOWN_TEXT` to print "this should be formatted like `code`", but instead would print "this should be formatted like \`code\`", erroneously keeping the backslashes in.

This commit teaches the environment parser to understand how to unescape backticks, fixing the issues laid out above.