Teach env.Environment to deal with multiple declare options when parsing output from export -p
Previously, if a line in the output of `export -p` contained multiple options on the `declare` call (eg, there was an array in there - `declare -ax ARRAY=("red", "green", "blue")`), env.FromExport() would have some very weird behaviour - it could eat variables entirely, leading to them being unset in the build that the agent ran - more context for buildkite staff [here](https://3.basecamp.com/3453178/buckets/23112918/messages/5089354108#__recording_5089608441), but the TL;DR is that a hook was exporting an array variable, and it was causing other variables - important ones like `PATH` and `PWD` - to go missing.

This commit teaches env.FromExport to be able to introspect the args passed to each `declare` call, and allows it to ignore anything that it won't be able to parse. At the moment, it ignores indexed arrays (`declare -a`), associative arrays (`declare -A`), integers (`declare -i`) and links (`declare -n`)