Provide option for base64 json trace context encoding
Back when I added trace propagation to the agent, I was short-sighted
and used the golang-binary format. Since the trace context was meant to
be propagated between builds of a pipeline, it was safe to assume the
consumers would always be able to decode golang-binary formats.

But now I want to trace additional spans within code that runs within a
Buildkite pipeline and link it to the parent trace of the pipeline
build. This will let me gather hollistic metrics about specific failure
modes and performance of various commands/tooling running in my repos.

So JSON was added, since it's a format all modern languages can easily
parse without the need for a special library.