Make windows hooks process env changes the same way as unix
This uses the new buildkite-agent env command introduced in [#1781](https://github.com/buildkite/agent/pull/1781), and uses it for hooks on windows machines as well. This change was scoped out of the initial fix, as windows machines weren't affected by the bug, and the effort of testing them would add time between the issue being reported and a fix being issued.

This PR ports the behaviour that we were previously using for hooks running on unix machines to windows machines as well, making the code a bunch nicer and allowing us to delete a whole bunch of useless env processing code.