Add git config protocol.file.allow=always to tests
This was disabled due to a [security vulnerability in git](https://github.blog/2022-10-18-git-security-vulnerabilities-announced/#cve-2022-39253).
The test TestCheckingOutLocalGitProjectWithSubmodules was only passing in CI because we run version of git in CI that did not receive the patch.

One solution would have been to set protocol.file.allow=always globally. However, that would mean that anyone who ran the tests would expose their
enviroment to that vulnerability, and permanently if the global config was not unset afterwards.