Use skopeo to copy image manifest list from ECR to dockerhub
And the images and their layers too.

Just pulling and pushing the image is not sufficient for multi-arch images.
Attempting to do so will just push the system's os/arch image to the second
registry. While `docker` command does expose the needed parts of the registry
API to copy a manifest list, `skopeo` and `crane` do. I found it much easier to
reuse the existing docker credentials with `skopeo` running as a docker
container than `crane`, so I went with it.