Restore old tini path
Tini used to be downloaded from github to /usr/sbin/tini. Now we install
it from the ubuntu repositories and it is installed to /usr/bin/tini.

Both are in the $PATH, so there should be no issue if a user invokes the
command as `tini`. Also, the default entrypoint is using the new path to
tini, so users starting the docker container without overriding the
default entrypoint should not have experienced any issues.

Fixes: #1928