Explictly cancel checkout context on job cancellation
When a job is canceled, the job runner interupts the bootstrap which
then calls the cancelation channel to cancel the rest of the job. When a
git checkout is in progress, the exit code from git was not -1 and
the checkout was then retried. When a job is canceled we should stop the
job, and not attempt to retry the checkout. This introduces a
cancelation context that is explictly checked and will not retry if the
job has been canceled.