Update cache and silence progress
When you're building docker images (in a Dockerfile), you don't want the
apk to be persisted to the Dockerfile, so the `--no-cache` flag is
appropriate. However, when you're in a container (as is the case here),
updating the cache once and reusing it between apk invocations is the
better pattern. In either case, the `--no-progress` flag is useful to
simplify the Buildkite job log, however.