Remove possibility of a nil map returned from ConfigurationToEnvironment
Previously we ignored errors from this method, but now we just log them.
Either way, execution continues, exposing us to the possibility of a nil
pointer dereference. Since go is not expressive enough to eliminate the
possibility of a nil pointer dereference, we have to be extra careful
to ensure we don't propagate nil pointer in this manner.