Downgrade deprecation message from Warning
The deprecation message isn't something that is actionable by people who
create builds (builders). It's only relevant if the plugin is depending on the
deprecated variable name, and the builder would not normally be aware of
this. Furthermore, it's often the case that the plugin will work fine
even though the environment variable name is "deprecated", so this does
not need to be a warning.