De-experimentify ansi-timestamps
Instead of enabling the current ANSI timestamp behaviour with no opt-out, this provides an escape hatch to the old defaults with --no-ansi-timestamps.

I thought for a bit about how to make ANSI timestamps and TimestampLines work together. It could avoid having a header times streamer for the purpose of streaming header times, but would still need the logic in the header times streamer to distinguish headers from non-headers.