Fix origin for mirrored submodules
In `defaultCheckoutPhase`, if there are mirrored submodules, the `repository` arg to `getOrUpdateMirrorDir` varies (per submodule). The clone in `updateGitMirror` uses the arg, but if the mirror exists and the commit is not present, the origin is replaced with `b.Repository`, pointing the mirror at the wrong remote before fetching.

Fixes a bug introduced in #1959