Lock library
The agent API exists not only to implement the lock subcommands, but to provide a service useful for agent internal locking purposes, build commands, hooks, and plugins. Some of that might be implemented in Go, therefore there should be a useful library available. 

While writing the library, I recalled that the previous implementation does not handle double-releasing a lock appropriately - suppose `lock release` is called manually as well as on an EXIT trap, while another process acquires the same lock. To fix this, the lock package generates a token stored in the lock when acquired, and must match in order to unlock. This is printed by `lock acquire` and is now a required argument to `lock release`.