Add ordered.Map
In the process of trying to parse pipelines with a greater semantic understanding (in order to do things with them, like sign command steps), but retain compatibility, I kept hitting the need for YAML unmarshaling into an *ordered* map. For example, even though the "env" block is essentially unordered, we step through it *in order* so that variables can interpolate into one another *in order*. Another example is plugin config: we reflect it to plugins both via env vars, but also as a JSON blob, and there's a fair chance some plugin is going to be (maybe even deliberately) relying on ordering within its config.

So I implemented an ordered map. It's the next evolution of yamltojson, in that it provides the same sort of order-preserving translation when you unmarshal YAML and marshal JSON, but you can also work with the contents of the map in a more map-like way. It's almost like the good old days of yaml.v2, but better!

Under the hood, ordered.Map is based on a slice of tuples, and an index (map from keys to slice indexes).

To be extra trendy, ordered.Map is generic. However, for now, unmarshaling requires string keys, and the only two uses are Map[string, string] and Map[string, any]. 

If you tell yaml.v3 to unmarshal into any, it will give you a map[string]any for mapping nodes. So any trickery is actually around co-opting UnmarshalYAML to prefer unmarshaling as a *Map[string,any] instead.

The unmarshalling tests are copied fairly closely from yamltojson, so I'm reasonably confident it's about as good as yamltojson is.