Replace ParserResult with Pipeline
`Pipeline` uses an `ordered.Map[string, string]` for `Env`, and a new
`ordered.Slice` for `Steps`. Unmarshaling into `ordered.Slice` prefers
`ordered.Map[string, any]` over `map[string]any`.

Why not `[]*ordered.Map[string, any]`? Because of steps like `"wait"`, which are
scalars.

`Pipeline` stores anything else in a `RemainingFields` which is an ordinary map.
I can't think of a reason to preserve ordering in fields that are not `Env` or
`Steps`, but if there is something, I would like to hear about it!

Plus a couple of related things:

* Interpolation is updated to work with the new types.
* Interpolating keys in the `Env` map is moved to `interpolateEnvBlock` and
  interpolated keys replace the old keys, rather than preserving uninterpolated keys into the output.
* Changed the parser API from a `Parser` struct taking fully-read input, to having `Parse` be a function consuming an`io.Reader` and an env map.
* Tweaked the tests to use the updated ordering for the top-level pipeline, and
  format the remarshaled JSON to make diffs easier to comb through.