Parse command and commands the same
The docs say `commands` is an alias for `command`. So one could write `command`’s value as multiple strings in a sequence, or write `commands`’ value as a single string.

Something that parses either single strings or sequences of strings into `[]string` is too generic for `pipeline`, so I put it in `ordered`.