Add Signature, Sign, and Verify
The main design consideration was how to get objects to describe their default set of fields to sign, and to fetch ~arbitrary values for verifying, because Signature includes the list of signed fields. The rest pretty much flows from that.

Signatures can be attached to command steps (to begin with) and cover just the command itself. The core piece (throwing fields and values at a hash) is the same between Sign and Verify, so I implemented both.

I made the algorithm interface big enough to handle both symmetric and asymmetric schemes, but have only implemented HMAC-SHA256 here.