artifact.<VerbAgent>.<VerbImperative> -> artifact.<VerbImperative>.Do
This commit changes the naming of actions in the artifact package and renames them. The previous formula was that the struct was named after the verb agent (eg, Downloader) and the primary action on that struct would be named after the verb imperative (eg. Download).

The new formula is that the struct is named after the imperative (eg Download) and the primary action is simply called `Do`.

The changes can be summarised:
old:
artifacts, err := artifact.NewSearcher(...).Search(...)

new:
artifacts, err := artifact.NewSearch(...).Do(...)

This commit also updates the artifact search gear to work in the same way as the other artifact actions - the struct takes a config, and the primary action is just `Do(ctx) ([]artifact, error)`