Fix parsing a duration as another duration
`JobStatusInterval` is already a duration. Casting it to a duration
again is a no-op. However, multiplying it by `time.Second` is not.

This is because the `time.Duration` type is an alias for `int64` and
represents nanoseconds. `time.Second` is therefore a constant 1_000_000_000.
Thus, multiplying `JobStatusInterval` by `time.Second` makes it a billion times longer.