Remove a double check for the existence of a local hook and log when it is missing in debug
I always have to do a double take when I read this code. But finally
found a reason to remove it! `Executor.hasLocalHook` calls
`Executor.localHookPath`, so in this method,
`Executor.executeLocalHook`, we are calling `Executor.localHookPath` twice.

This PR removes that and adds some logging to the job logs for when the hook
is not found (in debug mode).