Modify olfactor to allow multiple smells
Since the replacer produces the longest match, if one of the smells
is a prefix of another, only the longer match will be produced by the
replacer. This means if we just used a hashtable to store the smelt
smells, we would only be able to match the longer smell.

To remedy this, we use a trie to store the smells that the olfactor
smelt. This has the consequence that we don't need to explicitly
store a prefix of the smell if we want to sniff for it too.