Add `--allowed-repositories` option
This allows the agent to only work on a defined set of repos, and will make the agent refuse the job if the repo sent down by the buildkite backend (via the `BUILDKITE_REPO` envar) isn't in that list.

This should protect users more from the possibility of buildkite being pwned, while being easier than implementing a pre-bootstrap hook.