Upgrade pre-installed packages in docker images
Sometimes, the upstream distro will release a package but we had to wait
until they made a new image before our images would get the new package.
This means the security scanners will report fixable vulnerabilities that a
rebuild would not fix. With this change, we just have to make a new
build to pull in the patched packages.