experiment: pty-raw avoids LF->CRLF mapping by setting PTY to raw mode
Otherwise the PTY driver in the kernel will map LF (\n) to CR,LF (\r\n),
and possibly do other unwanted processing of the output stream.

This is behind an experiment because it's possible that by entering raw
mode we're turning off some default PTY input or output processing that
we actually rely on in some situations.