Only expose the template type to the public in the hooks package
We need to pass in something about the template to the public method
WriteHookTemplate, so that it knows which template to write.
But that does not need to be the template itself. Now, its just an indication
of which template to use, and the actual templates are private to the package.