Fix data race on Process.pid
The process can start, run, and output "Ready" before `p.pid` is set. Use of `p.pid` in `Interrupt` and `Terminate` is guarded by `p.mu`, but setting `p.pid` was not guarded. Even with the guard, it is still possible that `Interrupt` is called before `p.pid` is set, so I restored `<-p.Started()` (that I removed in #1852).