Remove hook wrapper generation tests
These are too fragile. We should tests that the hook wrapper does what
is expected, not that the text of the hook wrapper is what is expected.
In particular, we will be replacing the agent binary called in the hook
wrapper with a dynamic value, so the tests will will call the same code
as they were compiled with. Otherwise, the tests will will call the
agent binary in the $PATH, which is unlikely to be the same version that
the tests were compiled with.