Change annotation priority default from 3 to 1
Currently annotation priority defaults are set server side, this means
that agent version doesn't currently control the default (even if they
mention it in the help text). This means that we're not able to pin a
default annotation priority to an agent version.

This updates the annotation priority to be set on the agent, not the
server. It also changes the default from 3 to 1 because priority of 3
has been considered to be confusing and not intuitive. This now matches
agent priority.

Agent version v3.65.0 will continue to use the server side default
(which will be updated to 1).