Silence Job API Log Group
The Job API prints a log group whenever it starts, which is usually once
per job, but can be more than on agent-stack-k8s. However, I think the
Job API should be transparent to the user: it is not that relevant to
them unless they are developing against it, which should the minority of
the time they are using buildkite. This removes the log group appearing in
the job logs unless debug is enabled.