Artifact creation: Break retry on any 4xx code
There's a debate around which 4xx code we're going to use. #2697 mentions 422 in the title but adds handling for 403. Some people thing 429 is better.

Rather than playing whack-a-mole with all the 4xx codes, I think we should do the same thing as the log uploader and break on any of them.