Remove the --debug-http flag from the git credential helper
one of the things that came up in a security chat the other day was the worry that when using buildkite hosted compute, using `--debug --debug-http` on the agent could get the agent’s git credential helper to print out the response body of the request asking for a code access token, and leak it into job logs.
i’ve verified that this probably can’t happen right now, for a couple of reasons:
- first, buildkite's compute layer controls the agents, and doesn’t allow users to set specific flags on the agent. malicious (or negligent) users wouldn’t be able to set the two flags that could cause the response body to be leaked into job logs
- regardless of this, even when those options are set, the output of the git credential helper is eaten by git itself, and so wouldn’t be visible to users anyway. you can see an example of this in action in [this job](https://buildkite.com/bennos-super-secret-cool-time-org/private/builds/46#018f5197-b602-4dd4-9e64-eb879f8c9534) (private) - you can see that the HTTP request and response for a metadata exists call are printed, but the details of the code access token request are not
regardless of this, i’m still keen to remove the --debug-http option from the agent’s git credential helper. there’s already multiple layers of defence here, but output that has the opportunity for misuse is still being produced, and if we don’t produce it, it can’t be misused.

to wit, this PR does just that - removes the `--debug-http` flag from the git credential helper command.