fix enable http/2 by default as intended by flags
At some point the agent HTTP client configuration was updated however this resulted in it falling back to HTTP/1.1. Ideally we would take advantage of HTTP/2 to improve overall performance, especially for agents who are connecting from locations outside the US.

This will also lead to less TLS handshakes, and connections overall, which in turn should reduce the likehood of stalls uploading artifacts.