Make it so the log streamer can be signaled to stop immediately
Currently if the agent is sent a sig term the log streamer will wait for all chunks to be uploaded, and even if a second sig term is sent we still wait.

Other components of the agent honor the a graceful=false flag which allows operators to force the agent to stop.

This change adds that graceful option to the log streamer stop.