Capture some HTTP client details from registering agents
Agent registration requests include an array of "features" that are
enabled on the agent and influence the agent behaviour. To assist with
debugging, it would be helpful to know if a particular agent:

* has opted into HTTP 1.1 only
* is using a HTTP proxy
* has set GODEBUG, which can be used to change the behaviour of net/http [1]

It would also be quite helpful to have the agent log out the state of
its connection pool to agent.buildkite.com, but that's quite a different
change to this so I've left it out and kept this PR simple.

[1] https://pkg.go.dev/net/http