api.Client.requestHeaders is http.Header not map[string]string
http.Header is map[string][]string… we don't really want to use multiple
values per key, but using this type here is a bit more idiomatic, and
not too awkward.

Whereas we continue to use map[string]string for the API response
structs, since the JSON will contain a single string value per key, not
an array of strings, so it can't be automatically deserialized into a
http.Headers map.