Ignore the ping interval if agent will disconnect after job
Since the change in 3.95.0 to ping after a job is finished we've had reports that jobs are assigned to agents that are about to be disconnected.
While this could happen before this change, we believe waiting for the ping interval has made it more frequent and it's likely to happen multiple times for the same job.

This change ensures we don't wait up to 10s before disconnecting, making it as likely as before