Print a better error when the user uses an unknown subcommand
Previously, it would just print the help text without any errors, now, it'll print `buildkite-agent: unknown subcommand "chartreuse".

Additionally, this commit makes it so that the agent, when run without any subcommands, will exit 0 instead of exiting 1. The feature that made us exit 1 in this case was overriding the unknown subcommand behaviour