Relax Go version directive
Without this change, dependents must specify the same Go version or
newer in their go.mod file. The go version directive only needs to be
the lowest version that provides the language features needed, which
would be 1.x.0, with dependents building their packages using the latest
toolchain.

With the updated buildkite/shellwords, we can relax the go version
directive to the minor version rather than a specific patch version.

Signed-off-by: Hayden <8418760+haydentherapper@users.noreply.github.com>