Fix beta pipeline OIDC claim names
In https://github.com/buildkite/agent/pull/3614, we added `branch_name` to the list of claims that we add to the OIDC token we use for interacting with AWS in the process of a beta release.

Unfortunately, we used the wrong claim name — it's `build_branch` not `branch_name`. This PR fixes that.